LPD-17036 Move constant file to the api module
